renamed TimelineStepLabel.js to TimelineStepHeading.js
